add tests for tick scaling (#17)
* add tests for tick scaling

* skip suplicate branch test when in pr

+++ glug_timer/src/frac.h
@@ -5,8 +5,8 @@
 
 typedef struct frac
 {
-    uint32_t numer;
-    uint32_t denom;
+    uint64_t numer;
+    uint64_t denom;
 } frac_t;
 
 #endif // GLUG_FRAC_H
